Moved components; update paths
