+++ frontend/src/components/Notice.jsx
